pyproject.toml
[project]
name = "personalfinance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.7",
    "django-extensions>=4.1",
    "django-money>=3.5.4",
    "django-tailwind>=4.2.0",
    "django-unfold>=0.69.0",
    "dotenv>=0.9.9",
]
